tests: Check for duplication in __local_fixups__ and __fixups__
A dts file that is over-determined in the sense that it already contains
both a __local_fixups__ and/or a __fixups__ node and phandles that
result in entries in these nodes, should not compiled to a device tree
that has duplicate entries. This was a problem before commit
915daadbb62d ("Start with empty __local_fixups__ and __fixups__ nodes").

Add a test that ensures this issue isn't reintroduced later.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@baylibre.com>